AUT-4198: Add cleanup hooks to API tests
Some of the API tests are using the UI to perform a complete
test but the sessions weren't being cleaned up correctly.

Updated Hooks to explicitly do the cleanup.  Moved the @After
methods into the Hooks class to make it easier to see what's
going on.
